Make location code a plugin and remove tests
